Tetris

Description

A simple Tetris implementation.

Documentation

data Game

A Tetris game's state.

newGame :: YX -> IO (Maybe Game)

Generates a new gave with the given extent. For example, a game with height 24 and width 10 (the standard) can be generated with newGame (YX 24 10).

activeCoords :: Game -> Set YX

Returns the active piece's coordinates.

frozenCoords :: Game -> Set YX

Returns the coordinates of all inactive pieces.

data Action

A user action.

Constructors

MoveDown

MoveLeft

MoveRight

Rotate

onAction :: Action -> Game -> Game

Updates the state of a game with a user's action. If the user's action is invalid (e.g. moving a piece out of bounds), the game's state will remain unchanged.

newtype Score

A thin wrapper around a game's score.

Constructors

Score Int

currentScore :: Game -> Score

Returns the number of full rows cleared so far.

onTick :: Game -> IO (Either Score Game)

Updates the state of the game with the passage of time. More precisely, the following actions are performed in order:

Attempts to move the active piece down. If this is a valid move, the function simply returns the updated game's state.

Freezes the active piece.

Clears all newly full rows, increasing the score and shifting other rows correspondingly.

Generates a new random piece, if this is not possible the game is over and the function returns the game's final score instead of the updated state.