Tetris.Board

Documentation

data Piece

A Tetris piece.

Pieces are generated via randomPiece.

randomPiece :: Int -> IO Piece

Generates a random piece in the given column.

movePiece :: YX -> Piece -> Piece

Moves a piece by the given 2D offset.

rotatePiece :: Piece -> Piece

Rotates a piece clockwise 90 degrees..

pieceCoords :: Piece -> Set YX

Returns the piece's coordinates.

data Board

A rectangular Tetris board.

newBoard :: YX -> Maybe Board

Creates a new board with the given position as extreme point. If at least one of the given position's fields is negative, the function will return Nothing.

boardCoords :: Board -> Set YX

Returns the coordinates of all populated positions inside the board.

boardWidth :: Board -> Int

Returns the number of columns in the board.

fitsOn :: Piece -> Board -> Bool

Checks whether the given piece fits on the board (i.e. within the boundaries of the board and doesn't overlap with any existing pieces).

freezePiece :: Piece -> Board -> (Int, Board)

Adds the piece to the board, clearing any newly filled rows. This function returns the number of cleared rows along with the updated board.